fix(circular dependency): extract localStorage helpers to a leaf module
  ## Context

  `cacheUtils.ts` and `reactiveVars/*` formed a benign-but-fragile import cycle:
  `cacheUtils → reactiveVars → authTokenVar → cacheUtils`. `authTokenVar.ts:9`
  calls `getItemFromLS(...)` at module-init, so the cycle is safe only when
  `cacheUtils.ts` happens to finish evaluating before that line is reached.
  Production survives because `init.ts` imports `reactiveVars` before `cacheUtils`,
  forcing the "lucky" order — but under Jest the entry order is inverted on several
  test paths, hitting a TDZ crash, and any new module-init LS read could trip it
  in prod too.

  ## Description

  - Move `getItemFromLS` / `setItemFromLS` / `removeItemFromLS` from
    `core/apolloClient/cacheUtils.ts` into a new leaf module
    `core/utils/localStorage.ts` (zero imports — the invariant that breaks the cycle).
  - Move `localStorageKeys.ts` from `core/constants/` to `core/utils/` to keep the
    LS helpers and their keys co-located.
  - Migrate every consumer (12 prod files + 6 tests) to import from the new paths.
    `cacheUtils.ts` keeps a local import for the helpers it uses internally and
    does NOT re-export them — every call site moves to the leaf module explicitly.
  - Drop the stale "// Mock cacheUtils to break circular dependency" workaround in
    `currentOrganizationVar.test.ts`; switch test mocks of `~/core/apolloClient`
    to `~/core/utils/localStorage` and adopt the `jest.fn()`-inside-factory +
    external cast pattern in the 4 callsite-tests that hit the new entry path.
