+++ src/core/apolloClient/reactiveVars/authTokenVar.ts
@@ -1,6 +1,6 @@
 import { makeVar } from '@apollo/client'
 
-import { getItemFromLS, setItemFromLS } from '~/core/apolloClient/cacheUtils'
+import { getItemFromLS, setItemFromLS } from '~/core/utils/localStorage'
 
 export const AUTH_TOKEN_LS_KEY = 'authToken'
 export const TMP_AUTH_TOKEN_LS_KEY = 'tmpAuthToken'
