+++ src/core/apolloClient/reactiveVars/customerPortalTokenVar.ts
@@ -1,6 +1,6 @@
 import { makeVar } from '@apollo/client'
 
-import { getItemFromLS, setItemFromLS } from '~/core/apolloClient/cacheUtils'
+import { getItemFromLS, setItemFromLS } from '~/core/utils/localStorage'
 
 export const CUSTOMER_PORTAL_TOKEN_LS_KEY = 'customerPortalToken'
