+++ src/core/apolloClient/reactiveVars/duplicatePlanVar.ts
@@ -1,7 +1,7 @@
 import { makeVar, useReactiveVar } from '@apollo/client'
 
-import { getItemFromLS, setItemFromLS } from '~/core/apolloClient/cacheUtils'
 import { FORM_TYPE_ENUM } from '~/core/constants/form'
+import { getItemFromLS, setItemFromLS } from '~/core/utils/localStorage'
 import { StatusTypeEnum } from '~/generated/graphql'
 
 export type PLAN_FORM_TYPE = keyof typeof FORM_TYPE_ENUM
