+++ src/pages/auth/LoginOkta.tsx
@@ -9,9 +9,10 @@ import { Alert } from '~/components/designSystem/Alert'
 import { Button } from '~/components/designSystem/Button'
 import { Typography } from '~/components/designSystem/Typography'
 import { TextInputField } from '~/components/form'
-import { hasDefinedGQLError, setItemFromLS } from '~/core/apolloClient'
-import { REDIRECT_AFTER_LOGIN_LS_KEY } from '~/core/constants/localStorageKeys'
+import { hasDefinedGQLError } from '~/core/apolloClient'
 import { LOGIN_ROUTE, useLocation } from '~/core/router'
+import { setItemFromLS } from '~/core/utils/localStorage'
+import { REDIRECT_AFTER_LOGIN_LS_KEY } from '~/core/utils/localStorageKeys'
 import { addValuesToUrlState } from '~/core/utils/urlUtils'
 import { LagoApiError, useFetchOktaAuthorizeUrlMutation } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
