+++ src/pages/dashboards/Dashboards.tsx
@@ -5,9 +5,10 @@ import { useEffect, useMemo, useRef } from 'react'
 
 import { GenericPlaceholder } from '~/components/designSystem/GenericPlaceholder'
 import { Typography } from '~/components/designSystem/Typography'
-import { envGlobalVar, getItemFromLS, removeItemFromLS, setItemFromLS } from '~/core/apolloClient'
-import { SUPERSET_FILTERS_LS_KEY_PREFIX } from '~/core/constants/localStorageKeys'
+import { envGlobalVar } from '~/core/apolloClient'
 import { FeatureFlags, isFeatureFlagActive } from '~/core/utils/featureFlags'
+import { getItemFromLS, removeItemFromLS, setItemFromLS } from '~/core/utils/localStorage'
+import { SUPERSET_FILTERS_LS_KEY_PREFIX } from '~/core/utils/localStorageKeys'
 import { encodeRison } from '~/core/utils/risonEncoder'
 import { extractNativeFilters } from '~/core/utils/supersetFilters'
 import { useSupersetDashboardsQuery } from '~/generated/graphql'
